# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SupportMessageInput {
  /**
     * @minLength 1
     * @maxLength 4000
     */
  text: string;
}

export type ClassificationResultCategory = typeof ClassificationResultCategory[keyof typeof ClassificationResultCategory];


export const ClassificationResultCategory = {
  billing: 'billing',
  bug: 'bug',
  feature: 'feature',
  other: 'other',
} as const;

export type ClassificationResultUrgency = typeof ClassificationResultUrgency[keyof typeof ClassificationResultUrgency];


export const ClassificationResultUrgency = {
  low: 'low',
  normal: 'normal',
  high: 'high',
} as const;

export interface ClassificationMetadata {
  provider: string;
  model: string;
  duration_ms: number;
  cache_hit: boolean;
  retries: number;
  /** @nullable */
  input_tokens: number | null;
  /** @nullable */
  output_tokens: number | null;
}

export interface ClassificationResult {
  category: ClassificationResultCategory;
  urgency: ClassificationResultUrgency;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /**
     * @minLength 1
     * @maxLength 240
     */
  reason: string;
  metadata: ClassificationMetadata;
}