# lib/api-zod/src/generated/types/classificationMetadata.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClassificationMetadata {
  provider: string;
  model: string;
  duration_ms: number;
  cache_hit: boolean;
  retries: number;
  /** @nullable */
  input_tokens: number | null;
  /** @nullable */
  output_tokens: number | null;
}