# lib/api-zod/src/generated/types/classificationResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClassificationMetadata } from './classificationMetadata';
import type { ClassificationResultCategory } from './classificationResultCategory';
import type { ClassificationResultUrgency } from './classificationResultUrgency';

export interface ClassificationResult {
  category: ClassificationResultCategory;
  urgency: ClassificationResultUrgency;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /**
     * @minLength 1
     * @maxLength 240
     */
  reason: string;
  metadata: ClassificationMetadata;
}